Add uniffi bindings
