+++ .github/workflows/test.yaml
@@ -20,5 +20,24 @@ jobs:
       - uses: actions-rust-lang/setup-rust-toolchain@v1
         with:
           rustflags: ""
-      - name: Tests
+      - name: REST regression tests
         run: cargo test -- --test-threads=1
+      - name: UniFFI SDK smoke tests
+        run: cargo test --features uniffi --lib uniffi_smoke_tests:: -- --test-threads=1
+
+  parity-tests:
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    strategy:
+      fail-fast: false
+      matrix:
+        test_name: [zero_amount_invoice, send_receive]
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: true
+      - uses: actions-rust-lang/setup-rust-toolchain@v1
+        with:
+          rustflags: ""
+      - name: Run parity test by name
+        run: cargo test ${{ matrix.test_name }} -- --test-threads=1
