+++ .gitignore
@@ -1,21 +1,26 @@
-# Test files
-/datacore
-/dataindex
-/dataldk0
-/dataldk1
-/dataldk2
-/tmp
+# Runtime/test state
+/datacore/
+/dataindex/
+/dataldk0/
+/dataldk1/
+/dataldk2/
+/tmp/
 
-# will have compiled files and executables
-/target
-
-# These are backup files generated by rustfmt
+# Rust build outputs
+/target/
+/bindings/uniffi-bindgen/target/
 **/*.rs.bk
 
-# intellij files
+# IDE/editor/OS files
 .idea/
 .vscode/
-
-# MaxOS files
 .DS_Store
 *.swp
+
+# Generated UniFFI artifacts (regenerated in CI/local scripts)
+/bindings/swift/
+/bindings/kotlin/
+/bindings/kotlin-android/org/
+/bindings/kotlin-android/jniLibs/
+/bindings/python/dist/
+/bindings/python/rgb_lightning_node.py
