+++ Cargo.toml
@@ -4,8 +4,16 @@ version = "0.1.0"
 edition = "2021"
 rust-version = "1.85.0"
 
+[lib]
+name = "rgb_lightning_node"
+crate-type = ["lib", "staticlib", "cdylib"]
+
 [[bin]]
 name = "rgb-lightning-node"
+path = "src/bin/rgb-lightning-node.rs"
+
+[features]
+default = []
 
 [dependencies]
 amplify = { version = "=4.8.1", default-features = false }
@@ -55,6 +63,7 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 typenum = "1.17.0"
 uuid = { version = "1.11.0", default-features = false, features = ["v4"] }
 walkdir = "2.5.0"
+uniffi = { version = "0.28.3", features = ["build"], optional = true }
 zip = { version = "2.2.0", default-features = false, features = ["time", "zstd"] }
 
 [dev-dependencies]
@@ -65,6 +74,9 @@ once_cell = "1.20.0"
 reqwest = { version = "0.12", default-features = false, features = ["json", "multipart", "native-tls", "stream"] }
 serial_test = "3.1.1"
 tracing-test = "0.2.5"
+
+[build-dependencies]
+uniffi = { version = "0.28.3", features = ["build"], optional = true }
 
 [patch.crates-io]
 lightning = { path = "./rust-lightning/lightning" }
