+++ bindings/uniffi-bindgen/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "uniffi-bindgen"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+uniffi = { version = "0.28.3", features = ["cli"] }
