+++ bindings/uniffi-bindgen/src/main.rs
@@ -1,0 +1,3 @@
+fn main() {
+    uniffi::uniffi_bindgen_main()
+}
