+++ src/bin/rgb-lightning-node.rs
@@ -1,0 +1,6 @@
+use anyhow::Result;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    rgb_lightning_node::run_daemon().await
+}
