+++ src/ffi/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "uniffi")]
+mod types;
+
+#[cfg(feature = "uniffi")]
+pub use types::*;
