+++ src/routes/node/mod.rs
@@ -1,0 +1,19 @@
+use super::*;
+
+mod read;
+mod write_admin;
+mod write_channels;
+mod write_open_channel;
+mod write_wallet;
+
+pub(crate) use read::{
+    address, btc_balance, check_indexer_url, check_proxy_endpoint, estimate_fee, get_channel_id,
+    list_channels, list_peers, list_transactions, list_transfers, list_unspents, network_info,
+    node_info, sign_message,
+};
+pub(crate) use write_admin::{
+    backup, change_password, init, lock, restore, revoke_token, shutdown, unlock,
+};
+pub(crate) use write_channels::{close_channel, connect_peer, disconnect_peer, send_onion_message};
+pub(crate) use write_open_channel::open_channel;
+pub(crate) use write_wallet::{create_utxos, fail_transfers, send_btc, sync};
