+++ uniffi-android.toml
@@ -1,0 +1,3 @@
+[bindings.kotlin]
+android = true
+android_cleaner = true
